refactor: resides logic code in packages/scripts
