+++ packages/scripts/types/aliasUtils.d.ts
@@ -1,0 +1,3 @@
+declare function readAndParseAlias(rawAliases: string[]): {};
+declare function handleWarningMissingAlias(): void;
+export { handleWarningMissingAlias, readAndParseAlias };
